economy: Make daily system more efficient
Rather than reseting dailies at midnight UTC, the current time is saved under the user's profile when they collect their daily. When attempting to collect a daily next time, the current time and the last collected time are compared to see if a day has passed since then or not. This is more memory efficient and prevents getting a "double daily" by collecting right before 00:00 UTC.