economy: Update existing profile with missing keys and default values if needed
This works well as future-proofing as any new keys added to the defaultProfile will be added to existing ones next time they are used. Also, if something were to a user's profile which caused keys to go missing, their profile would be fixed with default values for whichever keys were missing.